Move SchemaQQ test types into separate module
