+++ test/Tests/SchemaQQ/TH.hs
@@ -1,19 +1,17 @@
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE QuasiQuotes #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE TypeApplications #-}
 
-module Tests.SchemaQQ.TH where
+module Tests.SchemaQQ.TH
+  ( module Tests.SchemaQQ.TH
+  , module Tests.SchemaQQ.Types
+  ) where
 
 import Language.Haskell.TH.Quote (QuasiQuoter(..))
 import Language.Haskell.TH.TestUtils (tryQErr')
 
 import Data.Aeson.Schema (schema)
 import Data.Aeson.Schema.Internal (ToSchemaObject, showSchemaType)
-
-type UserSchema = [schema| { name: Text } |]
-type ExtraSchema = [schema| { extra: Text } |]
-type ExtraSchema2 = [schema| { extra: Maybe Text } |]
+import Tests.SchemaQQ.Types
 
 -- | A quasiquoter for generating the string representation of a schema.
 schemaRep :: QuasiQuoter
